Move simulation func from nonspatial to sim module
